Reduce coupling and improve encapsulation
`accumulator<T, bool>` is now basically a PLAIN integrator that does not
keep track of `calls` and returns a `plain_result<T>`. All publicly
accessible integration routines wrap `accumulator` and convert
`plain_result` into their result types with additional information (such
as channel weights for the multi channel integrator). The accumulator
does not keep track of `calls` because the integrator may decide to not
call the `invoke` member function for performance. This currently only
happens in `multi_channel` which skips calling `invoke` if it knows that
the integrand evaluates to zero.